+++ lib/src/plural.dart
@@ -3,12 +3,17 @@ library inflection.plural;
 import 'dart:convert';
 
 import 'uncountable_nouns.dart';
+import 'irregular_plural_nouns.dart';
 
 class PluralEncoder extends Converter<String, String> {
-  List<List> _pluralInflectionRules = [];
+  final List<List> _inflectionRules = [];
 
   PluralEncoder() {
-    addPluralInflectionRules([
+    irregularPluralNouns.forEach((singular, plural) {
+      addIrregularInflectionRule(singular, plural);
+    });
+
+    [
       [r'$', (m) => 's'],
       [r's$', (m) => 's'],
       [r'^(ax|test)is$', (m) => '${m[1]}es'],
@@ -29,13 +34,28 @@ class PluralEncoder extends Converter<String, String> {
       [r'^(ox)$', (m) => '${m[1]}en'],
       [r'^(oxen)$', (m) => m[1]],
       [r'(quiz)$', (m) => '${m[1]}zes']
-    ].reversed);
+    ].reversed.forEach((rule) => addInflectionRule(rule.first, rule.last));
   }
 
-  void addPluralInflectionRules(Iterable<List> rules) {
-    rules.forEach((r) {
-      _pluralInflectionRules.add([new RegExp(r.first, caseSensitive: false), r.last]);
-    });
+  void addInflectionRule(String singular, dynamic plural) {
+    _inflectionRules.add([new RegExp(singular, caseSensitive: false), plural]);
+  }
+
+  void addIrregularInflectionRule(String singular, String plural) {
+    final s0 = singular.substring(0, 1);
+    final srest = singular.substring(1);
+    final p0 = plural.substring(0, 1);
+    final prest = plural.substring(1);
+
+    if (s0.toUpperCase() == p0.toUpperCase()) {
+      addInflectionRule('(${s0})${srest}\$', (m) => '${m[1]}${prest}');
+      addInflectionRule('(${p0})${prest}\$', (m) => '${m[1]}${prest}');
+    } else {
+      addInflectionRule('${s0.toUpperCase()}(?i)${srest}\$', (m) => '${p0.toUpperCase()}${prest}');
+      addInflectionRule('${s0.toLowerCase()}(?i)${srest}\$', (m) => '${p0.toUpperCase()}${prest}');
+      addInflectionRule('${p0.toUpperCase()}(?i)${prest}\$', (m) => '${p0.toUpperCase()}${prest}');
+      addInflectionRule('${p0.toLowerCase()}(?i)${prest}\$', (m) => '${p0.toLowerCase()}${prest}');
+    }
   }
 
   @override
@@ -44,7 +64,7 @@ class PluralEncoder extends Converter<String, String> {
       if (uncountableNouns.contains(word.toLowerCase())) {
         return word;
       } else {
-        for (var r in _pluralInflectionRules) {
+        for (var r in _inflectionRules) {
           RegExp pattern = r.first;
           if (pattern.hasMatch(word)) {
             return word.replaceAllMapped(pattern, r.last);
@@ -57,4 +77,4 @@ class PluralEncoder extends Converter<String, String> {
   }
 }
 
-final PLURAL = new PluralEncoder();
+final Converter<String, String> PLURAL = new PluralEncoder();