+++ lib/src/spinal_case.dart
@@ -5,7 +5,7 @@ import 'dart:convert';
 final _UNDERSCODE_RE0 = new RegExp(r'''([A-Z\d]+)([A-Z][a-z])''');
 final _UNDERSCODE_RE1 = new RegExp(r'''([a-z\d])([A-Z])''');
 
-class SpinalCaseEncoder extends Converter {
+class SpinalCaseEncoder extends Converter<String, String> {
   const SpinalCaseEncoder();
 
   /// Converts the input [phrase] to 'spinal case', i.e. a hyphen-delimited,
@@ -20,4 +20,4 @@ class SpinalCaseEncoder extends Converter {
   }
 }
 
-const SPINAL_CASE = const SpinalCaseEncoder();
+const Converter<String, String> SPINAL_CASE = const SpinalCaseEncoder();